Add files via upload
-Run this script in TERMUX. 
-After downloading you may need to use the command 'chmod u+x droid.sh' in order to make the file executable. Otherwise you run this by using the command 'bash droid.sh'.


This file is for android devices to get the miner set up and running. It includes adding commands to be used later on to restart the miner if you close it. 

-These commands are: 1. "STARTMINER" - used to begin mining again. 2. "VIEWMINER" - used to re-attach to a miner that you've previously detached from. 

This script includes a way to run the miner in a manner that can be detached from and ran in the background. That way you can close termux and have the miner continue on.