Add files via upload
This works for any debian based systems. 
Note this is not a wallet itself, but simply a fast way to register one that you can take with you to any DERO wallet. 

The output of this script will be in a file called KEEP_SAFE.txt and we highly recommend doing just that.

Script used is by https://github.com/deroholic/fastreg and a huge shoutout and thank you to them.